Switch to using wpt-metadata for triaged state.
This means that instead of looking at the gecko metadata for bug
annotations, we look in wpt.fyi. That's better because we can query
the data directly rather than have a post-hoc filter.

This change doesn't update the filter UI in general and in particular
the custom filtering parts don't allow expressing conditions that are
passed down to the server, rather than post-hoc conditions. That UI
needs to be rethought in a later commit.